+++ backend/src/main/java/com/ssafy/api/model/MyTrophy.java
@@ -23,7 +23,7 @@ import lombok.ToString;
 public class MyTrophy {
 	@Id
 	@Column(name = "USER_ID")
-	private int userID;
+	private long userID;
 
 	@Column(name = "TROPHY_ID")
 	private int trophyID;
